% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{Abies_ba_ratio_midpoint}
\alias{Abies_ba_ratio_midpoint}
\title{Abies_ba_ratio_midpoint}
\usage{
Abies_ba_ratio_midpoint()
}
\description{
Abies_ba_ratio_midpoint
}
\examples{
Abies_ba_ratio_midpoint()

}
