% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{Abies_mortality_ratio_midpoint}
\alias{Abies_mortality_ratio_midpoint}
\title{Abies_mortality_ratio_midpoint}
\usage{
Abies_mortality_ratio_midpoint()
}
\description{
Abies_mortality_ratio_midpoint
}
\examples{
mortality<-unlist(Abies_mortality_ratio_midpoint())
BAratio<-unlist(Abies_ba_ratio_midpoint())
TemperatureWIAbies_Population_BA_Mortality<-data.frame(TemperatureWIAbiesPopulation,BAratio,mortality)
# save(TemperatureWIAbies_Population_BA_Mortality,file="data/TemperatureWIAbies_Population_BA_Mortality.RData")
}
