% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig.R
\name{BA_calc.old}
\alias{BA_calc.old}
\title{Calculates basal areas of each period}
\usage{
BA_calc.old(d, sp = "")
}
\arguments{
\item{d}{forest data.}

\item{sp}{Roman species name. if default is sp="", that operate for all species}
}
\value{
　vector of sum of basal area for each period(year)
}
\description{
omit tree data with is.na(dbh) and dbh < 10cm
}
\examples{
plt
plot.<-"Matsuotoge"
d<-subset(dd3,plot==plot.) #d<-subset(dd3.,pn==6) #
BA_calc(d,"オオシラビソ")
(ba.<-BA_calc(d,""))
ba./ba.[1]
}
