% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig.R
\name{BA_calc}
\alias{BA_calc}
\title{Calculates basal areas of each period}
\usage{
BA_calc(d, sp = "スギ", dbh.min = 10, f.min = 1)
}
\arguments{
\item{d}{forest data.}

\item{sp}{Roman species name. if default is sp="", that operate for all species}

\item{dbh.min}{a minimum value of DBH for calculation (default = 10)}

\item{f.min}{a minimum value of vital index for calculation  (default = 1)}
}
\value{
　vector of sum of basal area for each period(year)
}
\description{
omit tree data with is.na(dbh) and dbh < 10cm
}
\examples{
plt
plot.<-"Kaminokodaira"
d<-subset(dd3,plot==plot.) #d<-subset(dd3.,pn==6) #
BA_calc(d,"スギ")
(ba.<-BA_calc(d,""))
ba./ba.[1]
plot.<-"Mimatsu"
d<-subset(dd3,plot==plot.)
(ba.<-BA_calc(d,"",dbh.min=0))
 (ba.<-BA_calc(d,"",dbh.min=10))
plot.<-"Kagamiishi"
d<-subset(dd3,plot==plot.)
plot(ba.<-BA_calc(d,"",dbh.min=0),type="b",ylim=c(0.4,1),main=plot.,ylab="Basal Area")
lines(ba..<-BA_calc(d,"",dbh.min=10),col="red",type="b")
legend(1.5,1.0,c("dbh>0cm","dbh>10cm"),col=c("black","red"),lty=c(1,1))

plot(ba./ba.[1],type="b",ylim=c(0.9,1.2),main=plot.)
lines(ba../ba..[1],type="b",col="red")
legend(1.5,1.15,c("dbh>0cm","dbh>10cm"),col=c("black","red"),lty=c(1,1))

}
