% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig.R
\name{BA_matrix2}
\alias{BA_matrix2}
\title{This makes Basal area matrix for 6 periods of tree species or All species}
\usage{
BA_matrix2(sp = "オオシラビソ", dbh.min = 10, f.min = 1)
}
\arguments{
\item{dbh.min}{a minimum value of DBH for calculation (default = 10)}

\item{f.min}{a minimum value of vital index for calculation  (default = 1)}

\item{d}{data frame of forest stand trees}
}
\description{
This makes Basal area matrix for 6 periods of tree species or All species
}
\examples{
plt
plot.<-"Kaminokodaira"
d<-subset(dd3,plot==plot.)

# Vital check ####
(f.<-d[,c(clm_f)])
table(as.numeric(as.matrix(f.)))
sum(is.na(f.))


# all tree species ####
sp.<- ""
(BA <- BA_matrix(sp.))
 (BA0 <- BA_matrix2(sp.,dbh.min=0))
# Kagamiishi trees with more 1.3m height, others trees with more 10cm dbh
BA[7,]<-BA0[7,]

# absolute ###
par(mfrow=c(1,1))
plot(0,type="n" ,xlim=c(1,6),ylim=c(0,90),
xlab="Period",ylab="Basal Area (m*m)")
for (ii in 1:8)lines(BA[ii,],type="b",lty=leg$lty[ii],col=leg$col[ii],pch=leg$pch[ii])

# ratio ####
(BAr <- BA/BA[,1])
rng<-range(BAr,na.rm=TRUE)
plot(0,type="n" , lty=ii,pch=ii,col=ii,
     xlim=c(0.25,6) , ylim=c(rng[1]-.02,rng[2]+.02),
     xlab="period",ylab="Basal area ratio",main=sp.)
abline(h=1)
plr <- which(BA[,1]>0) # plr : plot.recorded
for (ii in plr)lines(BAr[ii,],type="b",
                     lty=leg$lty[ii],pch=leg$pch[ii],col=leg$col[ii])
legend(1.2,1.2,leg$n,pch=leg$pch,col=leg$col,lty=leg$lty,cex=0.7)



# "ooshirabiso" ####
sp.<- "オオシラビソ"
(BA <- BA_matrix(sp.))
(BA0 <- BA_matrix2(sp.,dbh.min=0))
#Kagamiishi trees with more 1.3m height, others trees with more 10cm dbh
BA[7,]<-BA0[7,]
# absolute ###
par(mfrow=c(1,1))
plot(0,type="n" ,xlim=c(1,6),ylim=c(0,20),
xlab="Period",ylab="Basal Area (m*m)")
for (ii in 1:8)lines(BA[ii,],type="b",lty=leg$lty[ii],col=leg$col[ii],pch=leg$pch[ii])

# ratio ####
(BAr <- BA/BA[,1])
rng<-range(BAr,na.rm=TRUE)
plot(0,type="n" , lty=ii,pch=ii,col=ii,
     xlim=c(1,6) , ylim=c(rng[1]-.01,rng[2]+.01),
     xlab="period",ylab="Basal area ratio",main=sp.)
abline(h=1)
plr <- which(BA[,1]>0) # plr : plot.recorded
for (ii in plr)lines(BAr[ii,],type="b",
                     lty=leg$lty[ii],pch=leg$pch[ii],col=leg$col[ii])
 i<-4:7
legend(1,rng[2]+.01,leg$n[i],pch=leg$pch[i],col=leg$col[i],lty=leg$lty[i],cex=0.7)


}
