% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{ba_wi_plot3}
\alias{ba_wi_plot3}
\title{ba_wi_plot3}
\usage{
ba_wi_plot3(wi., d2)
}
\description{
ba_wi_plot3
}
\examples{

d<-TemperatureWIAbiesPopulation
wi.<-subset(d,plot=="Kaminokodaira")$WI
par(mfrow = c(2, 3))
d2<-.data_Fig_yr_ba_kaminokodaira_Cryptomeria_Fagus_Abies_2024
ba_wi_plot3(wi.,d2)
d2<-.data_Fig_yr_ba_kaminokodaira_zone_2024
ba_wi_plot3(wi.,d2)



}
