% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WI.R
\name{ci_calc}
\alias{ci_calc}
\title{Title}
\usage{
ci_calc(T12 = 1:12)
}
\arguments{
\item{t12}{}
}
\description{
Title
}
\examples{
ci_calc()
(t. <- plt5[,paste0("t",1:12)])
ci_calc(t.[1,])
apply(t.,1,wi_calc)
apply(t.,1,ci_calc)
}
