% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig.R
\name{dbh_hist}
\alias{dbh_hist}
\title{histgram of dbh including dead standing trees}
\usage{
dbh_hist(
  plotname = "Kaminokodaira",
  species = "オオシラビソ",
  term = 1,
  dbh_min = 10,
  breaks = seq(0, 120, 10),
  ...
)
}
\description{
histgram of dbh including dead standing trees
}
\examples{
par(mfrow=c(2,3))
for(i in 1:6){
dbh_hist("Kaminokodaira","オオシラビソ",i, 10,seq(0,60,5),ylim=c(0,15))
}

for(i in 1:6){
dbh_hist("Kagamiishi","オオシラビソ",i,0,seq(0,30,2),ylim=c(0,40))
}

}
