% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig.R
\name{death_ratio}
\alias{death_ratio}
\title{Title}
\usage{
death_ratio(
  data = subset(dd4, plot == "Kaminokodaira" & sp == "オオシラビソ"),
  form = paste0("f0", 1:7)
)
}
\description{
Title
}
\examples{
death_ratio()
years<-as.numeric(subset(plt2,na=="Kaminokodaira",(paste0("yr",1:7))))

 plot( years,death_ratio(data=subset(dd4,plot=="Kaminokodaira" & sp=="オオシラビソ"),form=paste0("f0",1:7)),type="b",
 ylab="枯死率",xlab="西暦年",main="初回2000年に生存していたオオシラビソ")


}
