% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationGrowth.R
\name{DemographicRates}
\alias{DemographicRates}
\title{DemographicRates  (Population growth : Condit  et al.)
  p <- log(n1/n0)/t *100
 m <- log(n0/nS)/t *100  Instantaneous Mortality Rate
 r <- log(n1/nS)/t *100}
\usage{
DemographicRates(n0 = 100, n1 = 150, nS = 80, t = 10)
}
\arguments{
\item{n0}{Population size at first survey}

\item{n1}{Population size at latest survey}

\item{nS}{Population size of survivors}

\item{t}{census intervals (years)}
}
\description{
DemographicRates  (Population growth : Condit  et al.)
  p <- log(n1/n0)/t *100
 m <- log(n0/nS)/t *100  Instantaneous Mortality Rate
 r <- log(n1/nS)/t *100
}
\examples{
DemographicRates(n0=100, n1=80, nS=50,t=24)
 # n0 n1 nS  t          p        m        r
 # 1 100 80 50 24 -0.9297648 2.888113 1.958348
 # m \% instantaneous mortality rate
 100*exp(-0.02888113*24)

DemographicRates(n0=100, n1=150, nS=80,t=10)
DemographicRates(n0=100, n1=100, nS=100,t=1)
DemographicRates(n0=100, n1=110, nS=100,t=1)

}
