% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{.pairwise_slope}
\alias{.pairwise_slope}
\title{2グループ間の傾き差をANCOVAで検定（内部関数）}
\usage{
.pairwise_slope(data, response, covariate, groupvar, ref, other)
}
\arguments{
\item{data}{data.frame}

\item{response}{目的変数名（文字列）}

\item{covariate}{共変量名（文字列）}

\item{groupvar}{グループ変数名（文字列）}

\item{ref}{参照グループ}

\item{other}{比較グループ}
}
\value{
named numeric vector
}
\description{
2グループ間の傾き差をANCOVAで検定（内部関数）
}
