% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WI.R
\name{F_kurobe_dum_mean_tmp}
\alias{F_kurobe_dum_mean_tmp}
\title{Return mean temperature from year with
 regression function between year and mean temperature in Kurobe dam}
\usage{
F_kurobe_dum_mean_tmp(year)
}
\value{
mean temperature
}
\description{
Return mean temperature from year with
 regression function between year and mean temperature in Kurobe dam
}
\examples{
kurobe_dam_temperature
kurobe_dam_temperature_regression
year<-1960:2023
plot(year,F_kurobe_dum_mean_tmp(year))
year00<-1953+14
abline(v=year00,col="red")  # base year of mesh data (1953-1982)

}
