% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig.R
\name{fallen_ratio}
\alias{fallen_ratio}
\title{オオシラビソの立枯木が倒伏する割合}
\usage{
fallen_ratio(
  d. = subset(dd4, plot == "Kaminokodaira" & sp == "オオシラビソ"),
  form = c("f03", "f04")
)
}
\arguments{
\item{d.}{a data.frame. The default is subset(dd4,plot=="Kaminokodaira" & sp=="オオシラビソ")}

\item{form}{a vector of character for form survey period, default is form=c("f03","f04")}
}
\description{
オオシラビソの立枯木が倒伏する割合
}
\examples{
# 立枯木の倒伏率の経年変化

fallen_ratio()
fallen_ratio(form="f07")
fallen_ratio(form=paste0("f0",1:7))
years<-as.numeric(subset(plt2,na=="Kaminokodaira",(paste0("yr",1:7))))
plot( years,fallen_ratio(form=paste0("f0",1:7)),type="b",
ylab="Falling ratio of standing dead trees",xlab="Year",main="A.mariesii in Kaminokodaira")


}
