% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WI.R
\name{Fig_alt_WI}
\alias{Fig_alt_WI}
\title{Fig_alt_WI}
\usage{
Fig_alt_WI()
}
\value{
not return anything, draw only.
}
\description{
Fig_alt_WI
}
\examples{
Fig_alt_WI()
ba.<-sp_zone_ba_ratio[[1]]
plt<-TateyamaForest2024$plot_profile
for(ii in 1:nrow(plt)){
  pichart(ba.[ii,],rx=40,ry=3,x=plt$alt[ii],y=plt$WI[ii],col=c("Orange","Purple","SkyBlue"))
}
text(plt$alt,plt$WI,plt$na,cex=0.8)

}
