% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WI.R
\name{Fig_alt_WI3}
\alias{Fig_alt_WI3}
\title{Fig_alt_WI3}
\usage{
Fig_alt_WI3(plt2, term = 1)
}
\arguments{
\item{term}{term of monitoring}
}
\value{
not return anything, draw only.
}
\description{
Fig_alt_WI3
}
\examples{
plt2<-TateyamaForest2024$plot_profile[-c(1,3,6,8),]
Fig_alt_WI2(plt2)
Fig_alt_WI3(plt3)

}
