% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WI.R
\name{Fig_alt_WI4_2020_JP}
\alias{Fig_alt_WI4_2020_JP}
\title{Fig_alt_WI4_2020_JP}
\usage{
Fig_alt_WI4_2020_JP(plt = plt5, term = 1)

Fig_alt_WI4_2020_JP(plt = plt5, term = 1)
}
\arguments{
\item{term}{term of monitoring}
}
\value{
not return anything, draw only.

not return anything, draw only.
}
\description{
Fig_alt_WI4_2020_JP

Fig_alt_WI4_2020_JP
}
\examples{
Fig_alt_WI4_2020_JP()


Fig_alt_WI4_2020_JP()


}
