% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WI.R
\name{Fig_alt_WI4_2020_Live_Dead}
\alias{Fig_alt_WI4_2020_Live_Dead}
\title{Fig_alt_WI4_2020_Live_Dead
　including standing dead trees}
\usage{
Fig_alt_WI4_2020_Live_Dead(plt = plt4, term = 1)
}
\arguments{
\item{term}{term of monitoring}
}
\value{
not return anything, draw only.
}
\description{
Fig_alt_WI4_2020_Live_Dead
　including standing dead trees
}
\examples{
Fig_alt_WI4_2020_Live_Dead()


}
