% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WI.R
\name{Fig_alt_WI4_2020_zone_2}
\alias{Fig_alt_WI4_2020_zone_2}
\title{Fig_alt_WI4_2020_zone_2()}
\usage{
Fig_alt_WI4_2020_zone_2(plt = plt5, term = 1)
}
\arguments{
\item{term}{term of monitoring}
}
\value{
not return anything, draw only.
}
\description{
Fig_alt_WI4_2020_zone_2()
}
\examples{
Fig_alt_WI4_2020_zone_2()


}
