% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WI.R
\name{Fig_alt_WI4_zone}
\alias{Fig_alt_WI4_zone}
\title{Fig_alt_WI4
　including standing dead trees}
\usage{
Fig_alt_WI4_zone(plt2, term = 1)
}
\arguments{
\item{term}{term of monitoring}
}
\value{
not return anything, draw only.
}
\description{
Fig_alt_WI4
　including standing dead trees
}
\examples{
Fig_alt_WI3(plt2,term=1)
Fig_alt_WI4(plt2,term=1)
　 Fig_alt_WI4_zone(plt2,term=1)

}
