% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig.R
\name{Fig_cunsumDeathRatio_2024}
\alias{Fig_cunsumDeathRatio_2024}
\title{Fig_cunsumDeathRatio_2024}
\usage{
Fig_cunsumDeathRatio_2024(
  sp. = "オオシラビソ",
  plot_name = c("Kaminokodaira", "Matsuotoge", "Kagamiishi"),
  main = "Abies mariesii (1999-2024)"
)
}
\value{
cunsumDeathRatio
}
\description{
Fig_cunsumDeathRatio_2024
}
\examples{
Fig_cunsumDeathRatio_2024()


}
