% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig.R
\name{Fig_Kaminokodaira_Abies_death_ratio}
\alias{Fig_Kaminokodaira_Abies_death_ratio}
\title{Title}
\usage{
Fig_Kaminokodaira_Abies_death_ratio()
}
\description{
Title
}
\examples{
Fig_Kaminokodaira_Abies_death_ratio()
}
