% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig.R
\name{Fig_live_standingdead}
\alias{Fig_live_standingdead}
\title{Make 4 dbh histgram of A. mariesii with live_standingdead}
\usage{
Fig_live_standingdead()
}
\description{
Make 4 dbh histgram of A. mariesii with live_standingdead
}
\examples{
Fig_live_standingdead()
 legend(5,30,c("Standing dead","Living"),col=c("gray","black"),pch=c(15,15))
}
