% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig.R
\name{Fig_live_standingdead_ratio}
\alias{Fig_live_standingdead_ratio}
\title{Fig_live_standingdead_ratio}
\usage{
Fig_live_standingdead_ratio()
}
\description{
Fig_live_standingdead_ratio
}
\examples{
Fig_live_standingdead_ratio()
}
