% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig2024.R
\name{Fig_snow_cover}
\alias{Fig_snow_cover}
\title{Fig_snow_cover}
\usage{
Fig_snow_cover(dat = snow)
}
\arguments{
\item{dat}{}
}
\description{
Fig_snow_cover
}
\examples{
Fig_snow_cover()

}
\seealso{
\code{\link{Fig_snow_depth}}
}
