% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig2024.R
\name{Fig_wi_ba_cor_ancova_JVS2}
\alias{Fig_wi_ba_cor_ancova_JVS2}
\title{Fig_wi_ba_cor_ancova_JVS2}
\usage{
Fig_wi_ba_cor_ancova_JVS2(clim_var = "WI")
}
\description{
Fig_wi_ba_cor_ancova_JVS2
}
\examples{
res<-Fig_wi_ba_cor_ancova_JVS2()
res
}
