% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig2024.R
\name{Fig_yr_ba_kaminokodaira_zone_2024}
\alias{Fig_yr_ba_kaminokodaira_zone_2024}
\title{Draw a figure relationships between year and basal area ratio for species main distribution zone in Kagamiishi in 2024}
\usage{
Fig_yr_ba_kaminokodaira_zone_2024()
}
\value{
data frame of this Figure data
}
\description{
Draw a figure relationships between year and basal area ratio for species main distribution zone in Kagamiishi in 2024
}
\examples{
Fig_yr_ba_kaminokodaira_zone_2024()
}
