% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig.R
\name{Fplt_yr_median}
\alias{Fplt_yr_median}
\title{Estimate death year of Abies using MTF (Mean Time to Fall) fall}
\usage{
Fplt_yr_median(site = "Kaminokodaira", period = 2:7)

Fplt_yr_median(site = "Kaminokodaira", period = 2:7)
}
\description{
Estimate death year of Abies using MTF (Mean Time to Fall) fall

Title
}
\examples{
names(dd4)
d. <- subset(dd4,plot=="Kaminokodaira" & sp=="オオシラビソ" & !is.na(d01))
i.<-d.$f01==0 & d.$f07==-1
sum(i.,na.rm=T)/sum(d.$f01==0,na.rm=T) # 初回立枯木25本のうち88\%の22本が7期までに倒れた
dbh.<-d.$d01[i.]
(mtf.<-MTF(dbh.,6.2)) #枯れながら立っていた時間
(period.<-apply(d.[i.,paste0("f0",1:7)]==-1,1,function(x)which(x)[1]))
#　何年に枯れたのか
(yr_fall.<-Fplt_yr_median("Kaminokodaira",period.)) #倒れた年
(yr_died. <- as.numeric(yr_fall.-mtf.))                        #枯れた年
h<-hist(yr_died.)
summary(yr_died.)
.<-yr_died.
m.<-round(mean(.),1)
sd.<-round(sd(.),1)
"The death years of dead standing Abies mariesii trees in 2000 at the Kaminokodaira plot were estimated to be 1990.9 ± 5.7 (mean ± s.d.) "
"using the MTF model (Gärtner et al., 2023), and their frequency distribution fitted a normal distribution (Kolmogorov-Smirnov test: D = 0.129, p = 0.811)."
paste(m.,"±", sd.,"(mean ± s.d.)")
sum(mean(.)-2*sd(.)<. & .<mean(.)+2*sd(.))/length(.)
ks.test(x=yr_died.,y="pnorm",mean=mean(yr_died.),sd=sd(yr_died.))


df <- data.frame(Year=years,Fallen_ratio=fallen_ratio(form=paste0("f0",1:7)))
plot( df,
      xlim=c(1972,2025),ylim=c(0,1),
      type="b",lwd=1.5,cex.lab=1.2,
      ylab="Falling ratio of dead standing trees",xlab="Year",main="Abies mariesii in Kaminokodaira",add=T)

polygon(data.frame(c(1972.5,h$mids,2012.5),1-0.88*c(0,10*h$density,0)),col="red",density=20)
text(2015,0.3,"Fall",cex=2)
rect(1987,0.8,1993,0.6,col="white",border="white") ; text(1990,0.7,"Death",cex=2,col="red")
names(dd4)
d. <- subset(dd4,plot=="Kaminokodaira" & sp=="オオシラビソ" & !is.na(d01))
i.<-d.$f01==0 & d.$f07==-1
sum(i.,na.rm=T)/sum(d.$f01==0,na.rm=T) # 初回立枯木25本のうち88\%の22本が7期までに倒れた
dbh.<-d.$d01[i.]
(mtf.<-MTF(dbh.,6.2)) #枯れながら立っていた時間
summary(mtf.)
paste("n=",length(mtf.)," mean=",mean(mtf.),"sd=",sd(mtf.),"min=",min(mtf.),"max=",max(mtf.))
(period.<-apply(d.[i.,paste0("f0",1:7)]==-1,1,function(x)which(x)[1]))
#　何年に枯れたのか
(yr_fall.<-Fplt_yr_median("Kaminokodaira",period.)) #倒れた年
(yr_died. <- as.numeric(yr_fall.-mtf.))                        #枯れた年
hist(yr_died.)
summary(yr_died.)
.<-yr_died.
m.<-round(mean(.),1)
sd.<-round(sd(.),1)
"The death years of dead standing Abies mariesii trees in 2000 at the Kaminokodaira plot were estimated to be 1990.9 ± 5.7 (mean ± s.d.) "
"using the MTF model (Gärtner et al., 2023), and their frequency distribution fitted a normal distribution (Kolmogorov-Smirnov test: D = 0.129, p = 0.811)."
paste(m.,"±", sd.,"(mean ± s.d.)")
sum(mean(.)-2*sd(.)<. & .<mean(.)+2*sd(.))/length(.)
ks.test(x=yr_died.,y="pnorm",mean=mean(yr_died.),sd=sd(yr_died.))

}
