% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WI.R
\name{LLToM3Code}
\alias{LLToM3Code}
\title{Tranlate mesh3 code from latitude nad longitude}
\usage{
LLToM3Code(lat = 36.63333, lon = 137.5875)
}
\value{
mesh3 code
}
\description{
Tranlate mesh3 code from latitude nad longitude
}
\examples{
sec <- 1/(360*60*60)

code. <- "54376430" #54377467"
 (ll<-M3CodeToLatLon(code.))  #メッシュの左下の座標
(code <- LLToM3Code(lat=ll[1], lon=ll[2]))
code == code.

#' メッシュ基点はメッシュの左下角
(code <- LLToM3Code(lat=ll[1]+10*sec, lon=ll[2]+10*sec))
code == code.
#' メッシュ基点からわずかに左下にずれた点は隣のメッシュ
(code <- LLToM3Code(lat=ll[1]-5*sec, lon=ll[2]-5*sec))
code == code.


plot(ll,xlim=c(ll[1]-60*sec,ll[1]+60*sec),ylim=c(ll[2]-60*sec,ll[2]+60*sec))
rect(ll[1],ll[2],ll[1]+45*sec,ll[2]+30*sec)
text(ll[1]+10*sec, ll[2]+10*sec,"in",col="blue")
text(ll[1]-5*sec, ll[2]-5*sec,"out !!",col="red")


# width and Height of mesh3 (45" 30" )

lat. <- 36
earth_radius <- 40000/(2*pi)
lat.length <- cos(36/180*pi)*earth_radius*2*pi
heigth.m3<-30*lat.length*sec
widthth.m3<-45*40000*sec
data.frame(heigth.m3,widthth.m3,arae.m3=heigth.m3 * widthth.m3)


LLToM3Code(plt$ido,plt$kei)


}
