% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{MannKendall_lm}
\alias{MannKendall_lm}
\title{Mann-Kendall test and lm regression analysis}
\usage{
MannKendall_lm(yr = wi_yaer$year, v = wi_yaer$Kaminokodaira)
}
\arguments{
\item{yr}{vector of years}

\item{v}{vector of values}
}
\description{
Mann-Kendall test and lm regression analysis
}
\examples{
MannKendall_lm(wi_year$year,wi_year$Kaminokodaira)

}
