% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationGrowth.R
\name{MortalityRate}
\alias{MortalityRate}
\title{discrete Mortality rate
 "discrete"   1 - (1 - p)^(1/d)
 "instantaneous"  log(1/p)/d}
\usage{
MortalityRate(p, d, method = c("instantaneous", "discrete"))
}
\arguments{
\item{p}{proportion of trees that DIED during the interval}

\item{d}{duration}

\item{method}{"instantaneous"(default),"discrete"}
}
\value{
Mortality Rate
}
\description{
discrete Mortality rate
 "discrete"   1 - (1 - p)^(1/d)
 "instantaneous"  log(1/p)/d
}
\examples{
 MortalityRate(0.48000000,24,method="instantaneous")
 # 0.03058205
 exp(-0.03058205*24)

Abies_death_ratio
# Kaminokodaira discrete
MortalityRate(0.50000000,24,method="discrete")
# 0.02687909
1-(1-0.02687909)^24

# Kaminokodaira "instantaneous"

MortalityRate(0.50000000,24)  #2000-2024
exp(-0.02888113*24)

MortalityRate(0.40000000,11)  #2013-2024 (0.50000000-0.10000000)
# 0.08329916
exp(- 0.08329916*11)

}
