% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig.R
\name{MTF}
\alias{MTF}
\title{MTF: Mean Time to Fall of Dead Standing Trees
Gärtner et al. (2023).
Temperature and Tree Size Explain the Mean Time to Fall of Dead Standing Trees across Large Scales.}
\usage{
MTF(DBH = 30, MAT = 5)
}
\arguments{
\item{MAT}{mean annual temperature}
}
\value{
Mean Time to Fall of Dead Standing Trees
}
\description{
MTF: Mean Time to Fall of Dead Standing Trees
Gärtner et al. (2023).
Temperature and Tree Size Explain the Mean Time to Fall of Dead Standing Trees across Large Scales.
}
\examples{
site.<-c("Bunazaka","Kaminokodaira","Matsuotoge","Kagamiishi" )
leg$col[match(site.,leg$n)]
col.<-c("darkolivegreen3","blueviolet","blue","cyan3" )
mat.<-c(8.0,6.2,2.8,1.0)
dbh.<-1:60
plot(dbh.,MTF(dbh.,mat.[1]),type="l",lty=1,col=col.[1],
xlab="DBH (cm)",ylab="Mean Time to Fall of Dead Standing Trees")
for(i in 2:4)lines(dbh.,MTF(dbh.,mat.[i]),col=col.[i],lty=i)
 legend(2,47,site.,col=col.,lwd=2,lty=1:4)



}
