% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig2024.R
\name{pichart}
\alias{pichart}
\title{Draw simple pi chart}
\usage{
pichart(X, rx = 1, ry = 1, x = 0, y = 0, col = 1:length(X), density = NA)
}
\arguments{
\item{X}{vector of pichart data}

\item{x}{x coordinate of center point}

\item{y}{y coordinate of center point}

\item{r}{radius of pichart circle}
}
\value{
no return values, only draw pichart
}
\description{
Draw simple pi chart
}
\examples{
X <-c(1,5,6)
plot(1:3)
pichart(X,rx=0.1,ry=0.1,x=2.5,y=2.5,col=c("Orange","Purple","SkyBlue"))
pichart(X,rx=0.1,ry=0.1,x=1.5,y=1.5,col=c("Orange","Purple","SkyBlue"),density =c(NA,20,10))



}
