% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationGrowth.R
\name{Population_growth}
\alias{Population_growth}
\title{Population growth (Condit  et al.)
save(PopulationGrowth,file="data/PopulationGrowth.RData")}
\usage{
Population_growth(plot. = "Kaminokodaira")
}
\arguments{
\item{plot.}{default = "Kaminokodaira".}
}
\value{
data.frame
}
\description{
Population growth (Condit  et al.)
save(PopulationGrowth,file="data/PopulationGrowth.RData")
}
\examples{
 Population_growth("Kaminokodaira")
 str(Population_growth("Kaminokodaira"))
  Population_growth("Kagamiishi")
  Population_growth("Bunazaka")
l<-lapply(plt2$na,Population_growth)
names(l)<-plt2$na
(PopulationGrowth <- l)

# save(PopulationGrowth ,file="data/PopulationGrowth.RData")

#' # Table作成 ####
spl<-SpeciesList5
l<-PopulationGrowth
l..<-c()
for(ii in 1:length(l)){
  l.<-l[[ii]]
  zone <- spl$zone[match(l.$sp,spl$spj)]
  l..<-rbind(l..,data.frame(plot=names(l)[ii],zone,l.))
}

(PopulationGrowth_df <-l..)
# save(PopulationGrowth_df ,file="data/PopulationGrowth_df.RData")
# write.csv(PopulationGrowth_df,file="data_raw/PopulationGrowth_df.csv")


l<-PopulationGrowth_df
plot.zone <- data.frame(
  plot= c("Bunazaka","Kaminokodaira","Kaminokodaira","Kaminokodaira","Matsuotoge","Kagamiishi"),
  zone= c("Temperate","Temperate","Ecotone","Subarctic","Subarctic","Subarctic")
)

plot.zone

l..<-c()
for(ii in 1:nrow(plot.zone)){
  l.<-subset(l,plot== plot.zone[ii,1] & zone==plot.zone[ii,2])
  nt<-c(sum(l.$n0),sum(l.$n1),sum(l.$nS),l.$t[1])
  l..<-rbind(l..,DemographicRates(nt[1],nt[2],nt[3],nt[4]))
}


(PopulationGrowth_zone_df<-data.frame(plot.zone,l..))

# save(PopulationGrowth_zone_df ,file="data/PopulationGrowth_zone_df.RData")

# write.csv(PopulationGrowth_zone_df ,file="data_raw/PopulationGrowth_zone_df.csv")

# 各プロット全種DemographicRates ####
l<-PopulationGrowth_zone_df
l$sp
l<-data.frame(cbind(tapply(l$n0,l$plot,sum),tapply(l$n1,l$plot,sum),tapply(l$nS,l$plot,sum)  ) ,c(22,24,24,23) )
names(l)<-c("n0","n1","nS","t")
unique(l$plot)

DemographicRates_all <- DemographicRates(l$n0,l$n1,l$nS,l$t)
rownames(DemographicRates_all)<-rownames(l)
DemographicRates_all

}
