% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{run_pairwise}
\alias{run_pairwise}
\title{全ペアの傾き比較（内部関数）}
\usage{
run_pairwise(data, response, covariate, groupvar)
}
\arguments{
\item{data}{data.frame}

\item{response}{目的変数名（文字列）}

\item{covariate}{共変量名（文字列）}

\item{groupvar}{グループ変数名（文字列）}
}
\value{
data.frame（結果表）
}
\description{
全ペアの傾き比較（内部関数）
}
