% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig2024.R
\name{sp_dominant_ba_ratio_calc}
\alias{sp_dominant_ba_ratio_calc}
\title{Return a data frame of total basal area ratio for each plot
 grouping with dominant tree species "スギ","ブナ","オオシラビソ"}
\usage{
sp_dominant_ba_ratio_calc()
}
\value{
a data frame of total basal area ratio
}
\description{
Return a data frame of total basal area ratio for each plot
 grouping with dominant tree species "スギ","ブナ","オオシラビソ"
}
\examples{
sp_dominant_ba_ratio_calc()
}
