% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig2024.R
\name{sp_zone_ba_calc}
\alias{sp_zone_ba_calc}
\title{Return a data frame of total basal area  for each plot}
\usage{
sp_zone_ba_calc(ba_list = sp_ba, term = 1)
}
\arguments{
\item{ba_list}{list of data frame for each plot}

\item{term}{term of monitoring}
}
\value{
a data frame
}
\description{
grouping with main distribution zone "Temperate", "Ecotone", "Subarctic"
}
\examples{
plt
sp_zone_ba_calc(ba_list=sp_ba,term=1)
sp_zone_ba<-c()
for (i in 1:7){
 sp_zone_ba<-c(sp_zone_ba,list(sp_zone_ba_calc(ba_list=sp_ba,term=i)))
}
sp_zone_ba
# save(sp_zone_ba,file="data/sp_zone_ba.RData")

sp_zone_ba_calc(ba_list=sp_ba_StandigDead,term=1)
sp_zone_ba_dead<-c()
for (i in 1:7){
 sp_zone_ba_dead<-c(sp_zone_ba_dead,list(sp_zone_ba_calc(ba_list=sp_ba_StandigDead,term=i)))
}
sp_zone_ba_dead
# save(sp_zone_ba_dead,file="data/sp_zone_ba_dead.RData")

}
