% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig2024.R
\name{sp_zone_ba_ratio_calc2}
\alias{sp_zone_ba_ratio_calc2}
\title{Return a data frame of total basal area ratio for each plot
(Living + Standing Dead)}
\usage{
sp_zone_ba_ratio_calc2(term = 1)
}
\arguments{
\item{term}{term of monitoring}
}
\value{
a data frame
}
\description{
grouping with main distribution zone "Temperate", "Ecotone", "Subarctic"
}
\examples{
plt
sp_zone_ba_ratio_calc(1)
sp_zone_ba_ratio_calc(7)
(sp_zone_ba_ratio<-lapply(1:7,sp_zone_ba_ratio_calc))
# save(sp_zone_ba_ratio,file="data/sp_zone_ba_ratio.RData")


}
