% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{statistics_ANCOVA_Abies_CumulativeDeathRatio_year_plot}
\alias{statistics_ANCOVA_Abies_CumulativeDeathRatio_year_plot}
\title{2. ANCOVA: 累積死亡率 ~ year * plot（傾きのペアワイズ比較）}
\usage{
statistics_ANCOVA_Abies_CumulativeDeathRatio_year_plot(s)
}
\arguments{
\item{s}{"statistics" オブジェクト}
}
\value{
list（model summary, slopes, pairwise comparisons）
}
\description{
2. ANCOVA: 累積死亡率 ~ year * plot（傾きのペアワイズ比較）
}
\examples{
 s <- new_statistics(
wi_year          = wi_year,
Abies_death_ratio = Abies_death_ratio,
f1_raw           = .data_Fig_yr_ba_kaminokodaira_Cryptomeria_Fagus_Abies_2024,
f2_raw           = .data_Fig_yr_ba_kaminokodaira_zone_2024
)

(res<-statistics_ANCOVA_Abies_CumulativeDeathRatio_year_plot(s))
str(res)

}
