% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{statistics_cor_wi_ba_EcotonePlot}
\alias{statistics_cor_wi_ba_EcotonePlot}
\title{3. Pearson相関: WI vs. BA比（樹種別・zone別、Kaminokodaira）}
\usage{
statistics_cor_wi_ba_EcotonePlot(s)
}
\arguments{
\item{s}{"statistics" オブジェクト}
}
\value{
list（by_species, by_zone）各data.frame（sp, r, p）
}
\description{
3. Pearson相関: WI vs. BA比（樹種別・zone別、Kaminokodaira）
}
\examples{
 s <- new_statistics(
wi_year          = wi_year,
Abies_death_ratio = Abies_death_ratio,
f1_raw           = .data_Fig_yr_ba_kaminokodaira_Cryptomeria_Fagus_Abies_2024,
f2_raw           = .data_Fig_yr_ba_kaminokodaira_zone_2024
)
(res<-statistics_cor_wi_ba_EcotonePlot(s))
str(res)

}
