% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationGrowth.R
\name{YearsIntervalCalc}
\alias{YearsIntervalCalc}
\title{YearsIntervalCalc}
\usage{
YearsIntervalCalc(yr, v, yr.interval, method = c("mean", "min", "max"))
}
\description{
YearsIntervalCalc
}
\examples{

# Dam_Tmin_max ####
d <- kurobe_dam_temperature
plot_name<-"Kagamiishi"
yr.interval<-as.numeric(subset(plt5,na==plot_name)[,paste0("yr",1:7)])
yr<-d$year
v<-d$max
YearsIntervalCalc(yr,v,yr.interval)
YearsIntervalCalc(yr,v,yr.interval,"mean")
YearsIntervalCalc(yr,v,yr.interval,"max")
YearsIntervalCalc(yr,v,yr.interval,"min")


}
