% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationGrowth.R
\name{YearsIntervalCalc_plots}
\alias{YearsIntervalCalc_plots}
\title{Title}
\usage{
YearsIntervalCalc_plots(
  plot.name = c("Kaminokodaira", "Matsuotoge", "Kagamiishi"),
  df = df,
  data.clm = "min",
  year.clm = "year",
  method = "min"
)
}
\arguments{
\item{method}{"mean","min","max"}
}
\description{
Title
}
\examples{

df <- kurobe_dam_temperature
str(df)
df1<-YearsIntervalCalc_plots(plot.name=c( "Kaminokodaira","Matsuotoge","Kagamiishi"),df=df,data.clm="min",year.clm="year",method="min")
df2<-YearsIntervalCalc_plots(plot.name=c( "Kaminokodaira","Matsuotoge","Kagamiishi"),df=df,data.clm="max",year.clm="year",method="max")
df3<-YearsIntervalCalc_plots(plot.name=c( "Kaminokodaira","Matsuotoge","Kagamiishi"),df=df,data.clm="mean",year.clm="year",method="mean")
d<-data.frame(df1,Tmax=df2$max,Tmean=df3$mean)
 names(d)[names(d)=="min"]<-"Tmin"
df.temp<-d
#標高補正　(dt<--0.55/100*(plt4$alt-1368))
 plot.name=c( "Kaminokodaira","Matsuotoge","Kagamiishi")
 for(ii in 1:3){
 plot.<-plot.name[ii]
 #dt.<- -0.55/100*(plt5$alt[plt5$na== plot.]-1368)
 dt.<- -0.55/100*(plt5$alt[plt5$na== plot.]-1459)
 df.temp$Tmin[df.temp$plot==plot.] <-df.temp$Tmin[df.temp$plot==plot.]+dt.
  df.temp$Tmax[df.temp$plot==plot.] <-df.temp$Tmax[df.temp$plot==plot.]+dt.
   df.temp$Tmaen[df.temp$plot==plot.] <-df.temp$Tmean[df.temp$plot==plot.]+dt.
 }

 # write.csv(df.temp,file="data_raw/YearsIntervalCalc_plots_temp.csv")


 # Warmth Index
 d<-YearsIntervalAverage_plots_wi_data
 d<-c(d$Kaminokodaira,d$Matsuotoge,d$Kagamiishi)
 TempWi<-data.frame(df.temp,wi=d)
  # write.csv(TempWi,file="data_raw/df_temp_wi.csv")

}
