% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig.R
\name{BasalArea_1_6}
\alias{BasalArea_1_6}
\title{returns basal area of a plot for a species}
\usage{
BasalArea_1_6(plotname = "Kaminokodaira", species = "ooshirabiso")
}
\arguments{
\item{species}{}
}
\value{
data.frame(Year,BasaAera)
}
\description{
returns basal area of a plot for a species
}
\examples{
plot. <- "Kaminokodaira"
sp.1 <- "ooshirabiso"
sp.2 <- "buna"
ba.sp1 <- BasalArea_1_6(plot.,sp.1)
ba.sp2 <- BasalArea_1_6(plot.,sp.2)
Year <- ba.sp1$Year
rba.sp1 <- 100*ba.sp1[,2]/ba.sp1[1,2]   #relative basal area
rba.sp2 <- 100*ba.sp2[,2]/ba.sp2[1,2]   #relative basal area
plot(Year,rba.sp1,ylab="Basal area (\%)", ylim=c(60,200),
type="b",lwd=5,pch=2,col="blue", main="Kaminokodaira (a.s.l. 1450m　ecotone)")
lines(Year,rba.sp2,type="b",col="orange",lwd=5)
abline(h=100,col="red",lty=2,lwd=2)
legend(2000,180,c("Abies mariesii","Fagus crenata"),lwd=5,pch=c(2,1),col=c("blue","orange"))
}
