% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbh_hist.R
\name{dbh_hist}
\alias{dbh_hist}
\title{histgram of dbh including dead standing trees}
\usage{
dbh_hist(plotname = "Kaminokodaira", species = "ooshirabiso", term = 1, ...)
}
\arguments{
\item{term}{}
}
\description{
histgram of dbh including dead standing trees
}
\examples{
par(mfrow=c(2,3))
for(i in 1:6){
dbh_hist("Kaminokodaira","ooshirabiso", i,ylim=c(0,13))
}


}
