% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig2024.R
\name{Df_table}
\alias{Df_table}
\title{summarize field stand data  to table composed of species(sp), DBH,vitality index}
\usage{
Df_table(plot.name = "Arimine")
}
\arguments{
\item{plot.name}{}
}
\value{
a data frame of plot species(sp), DBH(D),vitality index(f)
}
\description{
summarize field stand data  to table composed of species(sp), DBH,vitality index
}
\examples{
Df_table("Kagamiishi")
(d.<-Df_table("Arimine"))


}
