% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig2024.R
\name{Fig_Abies_DBH_hist_JVS2}
\alias{Fig_Abies_DBH_hist_JVS2}
\title{Distributions of the diameter at breast height of A. mariesii during the first survey (2000) in each plot.}
\usage{
Fig_Abies_DBH_hist_JVS2()
}
\description{
Distributions of the diameter at breast height of A. mariesii during the first survey (2000) in each plot.
}
\examples{
Fig_Abies_DBH_hist_JVS2()

}
