% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig2024.R
\name{Fig_Abies_wi_ba_mortality}
\alias{Fig_Abies_wi_ba_mortality}
\title{Fig_Abies_wi_ba_mortality}
\usage{
Fig_Abies_wi_ba_mortality()
}
\description{
Fig_Abies_wi_ba_mortality
}
\examples{
res<-Fig_Abies_wi_ba_mortality()
res
}
