% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WI.R
\name{Fig_kurobe_dam_temperature}
\alias{Fig_kurobe_dam_temperature}
\title{Title}
\usage{
Fig_kurobe_dam_temperature(d = kurobe_dam_temperature)
}
\description{
Title
}
\examples{
res<-Fig_kurobe_dam_temperature()
res

}
