% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WI.R
\name{Fig_year_WI_3}
\alias{Fig_year_WI_3}
\title{Fig_year_WI_3}
\usage{
Fig_year_WI_3()
}
\description{
Fig_year_WI_3
}
\section{Final TIFF output (JVS submission)}{

\preformatted{
tiff("Fig4.tiff", width = 3000, height = 2500, res = 300, compression = "lzw")
Fig_year_WI_3()
dev.off()
}
}

\examples{
 res<-Fig_year_WI_3()
 res
}
