% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig2024.R
\name{Fig_yr_ba_kaminokodaira_Cryptomeria_Fagus_Abies_2024}
\alias{Fig_yr_ba_kaminokodaira_Cryptomeria_Fagus_Abies_2024}
\title{Draw a figure relatinships between year and basal area ratio for Fagus - Abies in Kagamiishi in 2024}
\usage{
Fig_yr_ba_kaminokodaira_Cryptomeria_Fagus_Abies_2024()
}
\value{
nothing, only output figure
}
\description{
Draw a figure relatinships between year and basal area ratio for Fagus - Abies in Kagamiishi in 2024
}
\examples{
Fig_yr_ba_kaminokodaira_Cryptomeria_Fagus_Abies_2024()
}
