% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig2024.R
\name{Fig_yr_ba_site2}
\alias{Fig_yr_ba_site2}
\title{Fig_yr_ba_site2}
\usage{
Fig_yr_ba_site2()
}
\value{
nothing, draw a figure only.
}
\description{
Fig_yr_ba_site2
}
\examples{

Fig_yr_ba_site2()

\preformatted{
tiff("Fig5.tiff", width = 2000, height = 1700, res = 300, compression = "lzw")
TateyamaForest::Fig_yr_ba_site2()
dev.off()
# FIGURE 5 | Changes in the total basal area in each plot over time.
}


}
