% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WI.R
\name{M3CodeToLatLon}
\alias{M3CodeToLatLon}
\title{Tranlate to latitude and longitude  from mesh3 code}
\usage{
M3CodeToLatLon(M3Code = 54374091)
}
\arguments{
\item{M3Code}{}
}
\value{
latitude and longitude
}
\description{
Tranlate to latitude and longitude  from mesh3 code
}
\examples{

M3CodeToLatLon(54377467)
n2s(M3CodeToLatLon(54377467))



}
