% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WI.R
\name{n2s}
\alias{n2s}
\title{ddd.dddd -> dd.mm.ss}
\usage{
n2s(d)
}
\arguments{
\item{d}{}
}
\value{
string degree, minutes,secant ddd.mmssss
}
\description{
ddd.dddd -> dd.mm.ss
}
\examples{
d=137.50
n2s(d)
s2n(n2s(d))
#  第7系原点 radian ########
(lat0 = s2n("36.00000")*degree)
(lon0 = s2n("137.100000")*degree)
}
