% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{statistics_cor_WI_AbiesCumulativeMortality}
\alias{statistics_cor_WI_AbiesCumulativeMortality}
\title{1. Pearson相関: WI vs. Abies 累積死亡率（各プロット・プール）}
\usage{
statistics_cor_WI_AbiesCumulativeMortality(s)
}
\arguments{
\item{s}{"statistics" オブジェクト}
}
\value{
data.frame（plot, r, p, n）
}
\description{
1. Pearson相関: WI vs. Abies 累積死亡率（各プロット・プール）
}
\examples{
 s <- new_statistics(
wi_year          = wi_year,
Abies_death_ratio = Abies_death_ratio,
f1_raw           = .data_Fig_yr_ba_kaminokodaira_Cryptomeria_Fagus_Abies_2024,
f2_raw           = .data_Fig_yr_ba_kaminokodaira_zone_2024
)

(res<-statistics_cor_WI_AbiesCumulativeMortality(s))
str(res)

}
