% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{Table_TemperatureWIAbiesPopulation_cor}
\alias{Table_TemperatureWIAbiesPopulation_cor}
\title{Table. Pearson correlation coefficients
         between climate variables and demographic rates
         of A. mariesii
入力: TemperatureWIAbiesPopulation.csv
出力: Table6_updated.csv
============================================================}
\usage{
Table_TemperatureWIAbiesPopulation_cor()
}
\value{
data frame for arranged table
}
\description{
Table. Pearson correlation coefficients
         between climate variables and demographic rates
         of A. mariesii
入力: TemperatureWIAbiesPopulation.csv
出力: Table6_updated.csv
============================================================
}
\examples{
tbl<-Table_TemperatureWIAbiesPopulation_cor()
tbl
#write.csv(tbl, "Table_TemperatureWIAbiesPopulation_cor.csv", row.names = FALSE)

}
