% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig.R
\name{vital_mortality}
\alias{vital_mortality}
\title{plot relationships vital index and mortality from term 2 to 6}
\usage{
vital_mortality(d., sp. = "")
}
\arguments{
\item{d}{}
}
\value{
matrix of mortality
}
\description{
plot relationships vital index and mortality from term 2 to 6
}
\examples{

par(mfrow=c(1,4))
vital_mortality(dd2)        # all species ####
vital_mortality(dd2,"スギ") #Cryptomeria japonica スギ　####
vital_mortality(dd2,"ブナ") # Fagus crenata ブナ　####
vital_mortality(dd2,"オオシラビソ") # Abies mariesii オオシラビソ　####

}
