% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationGrowth.R
\name{YearsIntervalAverage}
\alias{YearsIntervalAverage}
\title{YearsIntervalAverage}
\usage{
YearsIntervalAverage(yr, v, yr.interval)
}
\arguments{
\item{yr.interval}{}
}
\description{
YearsIntervalAverage
}
\examples{
plot_name<-"Kagamiishi"
yr.interval<-as.numeric(subset(plt5,na==plot_name)[,paste0("yr",1:7)])
yr<-as.numeric(rownames(wi_year))
v<-as.numeric(wi_year[,plot_name])
YearsIntervalAverage(yr,v,yr.interval)
YearsIntervalCalc(yr,v,yr.interval,"mean")
}
