% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationGrowth.R
\name{YearsIntervalAverage_plots_wi}
\alias{YearsIntervalAverage_plots_wi}
\title{YearsIntervalAverage_plots}
\usage{
YearsIntervalAverage_plots_wi(
  plot.name = c("Kaminokodaira", "Matsuotoge", "Kagamiishi"),
  df = df,
  year.clm = "year"
)
}
\arguments{
\item{data}{}
}
\description{
YearsIntervalAverage_plots
}
\examples{
plot.name <- c( "Kaminokodaira","Matsuotoge","Kagamiishi")
df<-data.frame(wi_year); df$year=as.numeric(rownames(wi_year))
d<-YearsIntervalAverage_plots_wi(plot.name=c( "Kaminokodaira","Matsuotoge","Kagamiishi"),df=df,year.clm="year")
YearsIntervalAverage_plots_wi_data<-d
# save(d,file="data_raw/YearsIntervalAverage_plots_wi_data.RData")
}
