% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig2024.R
\name{yr_ba_site}
\alias{yr_ba_site}
\title{Return a list of data frame for year and basal area  per every site}
\usage{
yr_ba_site(cond = substitute(f. > 0))
}
\value{
list of data frame for year and basal area  per every site
}
\description{
Return a list of data frame for year and basal area  per every site
}
\examples{
BA <- yr_ba_site(cond=substitute(f.>0))
 plot(0,type="n" , lty=1,pch=1,col=1,
     xlim=c(1998,2025) , ylim=c(0.97,1.23),
     xlab="Year",ylab="Basal area ratio")
abline(h=1)
for (ii in 1:length(BA)){
 . <- BA[[ii]]
 lines(.$Year,.$BasalArea/.$BasalArea[1],type="b",
                     lty=leg$lty[ii],pch=leg$pch[ii],col=leg$col[ii])#'
}

legend(2000,1.2,leg$n,pch=leg$pch,col=leg$col,lty=leg$lty,cex=0.7)


}
