% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fig2024.R
\name{yr_sp_ba_site}
\alias{yr_sp_ba_site}
\title{Species basal area for each plot}
\usage{
yr_sp_ba_site(cond = substitute(f. > 0))
}
\arguments{
\item{cond}{condition of vital index(f.), default is "f.>0"(living tree)}
}
\value{
list of species basal area for each plot
}
\description{
save as "sp_ba.RData"
}
\examples{
(sp_ba <- yr_sp_ba_site())
(sp_ba_StandigDead <- yr_sp_ba_site(cond=substitute(f.==0)))

# save(sp_ba,file="data/sp_ba.RData")
# save(sp_ba_StandigDead,file="data/ssp_ba_StandigDead.RData")

(sp_ba_ratio<-sapply(sp_ba,function(x)t(t(x)/rowSums(t(x)))))
(sp_ba_ratio_StandigDead<-sapply(sp_ba_StandigDead,function(x)t(t(x)/rowSums(t(x)))))

# save(sp_ba_ratio,file="data/sp_ba_ratio.RData")
# save(sp_ba_ratio_StandigDead,file="data/sp_ba_ratio_StandigDead.RData")


}
